update from master

+++ blog/management/commands/build_index.py
@@ -1,20 +1,7 @@
-#!/usr/bin/env python
-# encoding: utf-8
-"""
-@version: ??
-@author: liangliangyy
-@license: MIT Licence 
-@contact: liangliangyy@gmail.com
-@site: https://www.lylinux.net/
-@software: PyCharm
-@file: build_index.py
-@time: 2019-04-20 20:39
-"""
-
-from blog.documents import ElapsedTimeDocument, ArticleDocumentManager
-
 from django.core.management.base import BaseCommand
-from blog.models import Article
+
+from blog.documents import ElapsedTimeDocument, ArticleDocumentManager, ElaspedTimeDocumentManager, \
+    ELASTICSEARCH_ENABLED
 
 
 # TODO 参数化
@@ -22,9 +9,10 @@ class Command(BaseCommand):
     help = 'build search index'
 
     def handle(self, *args, **options):
-        manager = ArticleDocumentManager()
-        manager.delete_index()
-        manager.rebuild()
-
-        manager = ElapsedTimeDocument()
-        manager.init()
+        if ELASTICSEARCH_ENABLED:
+            ElaspedTimeDocumentManager.build_index()
+            manager = ElapsedTimeDocument()
+            manager.init()
+            manager = ArticleDocumentManager()
+            manager.delete_index()
+            manager.rebuild()
